add get_footer call to front-page but hide the footer section with CSS
We need the 'get_footer()' call for WP to work properly - there is a lot that uses that hook. You probably noticed that you cannot see your admin bar anymore on your front page - this is part of the reason why. I added it back - wrapped your widget and colophon divs in a .footer section in footer.php and used it to hide it on the front page with your style.css